Переделал архитектуру, теперь летает не игрок, а его корабль (давно пора)

+++ SkyTrekVisual/GameItems/IDestructive.cs
@@ -8,7 +8,7 @@ namespace SkyTrekVisual.GameItems
 {
 	public interface IDestructive
 	{
-		int Damage { get; set; }
+		double Damage { get; set; }
 
 	}
